enterprise-doc-intelligence/requirements.txt
pypdf
python-docx
sentence-transformers
torch
faiss-cpu
fastapi
uvicorn


